Include symlinks during zip
Necessary because of an issue encountered when using resignme on an IPA
that has been packaged with symlinks (apparently).

http://stackoverflow.com/a/7016017

This shouldn't be the responsibility of fsutils zip - if anything it
should accept zip args as function parms. But this can be updated in the
future, for now this is sufficient.